doc: RDN values and create-only attributes
Add notes which fields are the primary keys / RDN values (relative
distinguised names) of objects. IPA does not support `modrdn` for
servers, certificates, and DNS locations. That means the entries cannot
be renamed and an RDN value identifies an entry.

Also note that domain name and realm name cannot be modified by an
update.

Signed-off-by: Christian Heimes <cheimes@redhat.com>