refactor(HMS-2191): remove metrics openapi
This change remove metrics.openapi.* files, and update Makefile according to
that.

This is a special case as the /openapi.json endpoint, and to achieve this to
work, the openapi file becomes invalid with the specification for the path of
the endpoint; be aware the /metrics path could differ and the port where it will
be listening, depending on the clowder configuration.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>