fix(HMS-3152): remove pattern matching from location name
Location name should be a DNS label and could be matched by a regex.

Correct one could be "^[a-zA-Z0-9]([a-zA-Z0-9\\-]*[a-zA-Z0-9])*$".

PR https://github.com/podengo-project/idmsvc-api/pull/21 has a
discussion about it. But users can defined more values in IPA
location in the FreeIPA server than what the RFCs allow.

From this PoV, it is more important for idmsvc to pass registration
even with slightly invalid DNS label than being valid but failing.
Idmsvc doesn't use the value for DNS operations.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>